cleanup(nx): clean up schematics to only depend on non-framework specific packages
