+++ packages/angular/src/schematics/library/library.spec.ts
@@ -1,7 +1,6 @@
 import { Tree, VirtualTree } from '@angular-devkit/schematics';
-import { createApp, createEmptyWorkspace } from '@nrwl/workspace/testing';
-import { runSchematic } from '../../utils/testing';
-import { getFileContent } from '@schematics/angular/utility/test';
+import { createEmptyWorkspace, getFileContent } from '@nrwl/workspace/testing';
+import { createApp, runSchematic } from '../../utils/testing';
 import * as stripJsonComments from 'strip-json-comments';
 import { NxJson, readJsonInTree } from '@nrwl/workspace';
 import { UnitTestTree } from '@angular-devkit/schematics/testing';
