+++ packages/angular/src/schematics/ngrx/ngrx.spec.ts
@@ -1,16 +1,18 @@
 import { UnitTestTree } from '@angular-devkit/schematics/testing';
 import { Tree, VirtualTree } from '@angular-devkit/schematics';
-import { getFileContent } from '@schematics/angular/utility/test';
 import { readJsonInTree } from '@nrwl/workspace';
 
 import { findModuleParent } from '@nrwl/workspace';
-import { AppConfig, getLibConfig, getAppConfig } from '@nrwl/workspace/testing';
-import { runSchematic } from '../../utils/testing';
+import { getFileContent } from '@nrwl/workspace/testing';
 import {
+  AppConfig,
   createApp,
   createLib,
-  createEmptyWorkspace
-} from '@nrwl/workspace/testing';
+  getAppConfig,
+  getLibConfig,
+  runSchematic
+} from '../../utils/testing';
+import { createEmptyWorkspace } from '@nrwl/workspace/testing';
 
 describe('ngrx', () => {
   let appTree: Tree;
