+++ packages/web/src/utils/normalize.ts
@@ -1,12 +1,6 @@
 import { WebBuildBuilderOptions } from '../builders/build/build.builder';
 import { Path, normalize } from '@angular-devkit/core';
 import { resolve, dirname, relative, basename } from 'path';
-import { BuildOptions } from '@angular-devkit/build-angular/src/angular-cli-files/models/build-options';
-import {
-  NormalizedBrowserBuilderSchema,
-  AssetPattern,
-  AssetPatternObject
-} from '@angular-devkit/build-angular';
 import { BuildBuilderOptions } from './types';
 import { statSync } from 'fs';
 
@@ -35,11 +29,7 @@ export function normalizeBuildOptions<T extends BuildBuilderOptions>(
   };
 }
 
-function normalizeAssets(
-  assets: AssetPattern[],
-  root: string,
-  sourceRoot: Path
-): AssetPatternObject[] {
+function normalizeAssets(assets: any[], root: string, sourceRoot: Path): any[] {
   return assets.map(asset => {
     if (typeof asset === 'string') {
       const assetPath = normalize(asset);
@@ -118,11 +108,9 @@ export function normalizeWebBuildOptions(
   };
 }
 
-export function convertBuildOptions(
-  buildOptions: WebBuildBuilderOptions
-): BuildOptions {
+export function convertBuildOptions(buildOptions: WebBuildBuilderOptions): any {
   const options = buildOptions as any;
-  return <NormalizedBrowserBuilderSchema>{
+  return <any>{
     ...options,
     buildOptimizer: options.optimization,
     aot: false,
