+++ packages/workspace/src/utils/ast-utils.ts
@@ -6,183 +6,189 @@
  * found in the LICENSE file at https://angular.io/license
  */
 import { Rule, Tree, SchematicContext } from '@angular-devkit/schematics';
-import {
-  findNodes,
-  getDecoratorMetadata,
-  getSourceNodes,
-  insertAfterLastOccurrence
-} from '@schematics/angular/utility/ast-utils';
-import {
-  Change,
-  InsertChange,
-  NoopChange,
-  RemoveChange,
-  ReplaceChange
-} from '@schematics/angular/utility/change';
-
 import * as ts from 'typescript';
-import * as path from 'path';
-import { toFileName } from './name-utils';
 import { serializeJson } from './fileutils';
-import * as stripJsonComments from 'strip-json-comments';
 import { NodePackageInstallTask } from '@angular-devkit/schematics/tasks';
 
-export function addReexport(
-  source: ts.SourceFile,
-  modulePath: string,
-  reexportedFileName: string,
-  token: string
-): Change[] {
-  const allExports = findNodes(source, ts.SyntaxKind.ExportDeclaration);
-  if (allExports.length > 0) {
-    const m = allExports.filter(
-      (e: ts.ExportDeclaration) =>
-        e.moduleSpecifier.getText(source).indexOf(reexportedFileName) > -1
-    );
-    if (m.length > 0) {
-      const mm: ts.ExportDeclaration = <any>m[0];
-      return [
-        new InsertChange(modulePath, mm.exportClause.end - 1, `, ${token} `)
-      ];
-    }
-  }
-  return [];
+function nodesByPosition(first: ts.Node, second: ts.Node): number {
+  return first.getStart() - second.getStart();
 }
 
-// This should be moved to @schematics/angular once it allows to pass custom expressions as providers
-function _addSymbolToNgModuleMetadata(
-  source: ts.SourceFile,
-  ngModulePath: string,
-  metadataField: string,
-  expression: string
-): Change[] {
-  const nodes = getDecoratorMetadata(source, 'NgModule', '@angular/core');
-  let node: any = nodes[0]; // tslint:disable-line:no-any
+function insertAfterLastOccurrence(
+  nodes: ts.Node[],
+  toInsert: string,
+  file: string,
+  fallbackPos: number,
+  syntaxKind?: ts.SyntaxKind
+): Change {
+  // sort() has a side effect, so make a copy so that we won't overwrite the parent's object.
+  let lastItem = [...nodes].sort(nodesByPosition).pop();
+  if (!lastItem) {
+    throw new Error();
+  }
+  if (syntaxKind) {
+    lastItem = findNodes(lastItem, syntaxKind)
+      .sort(nodesByPosition)
+      .pop();
+  }
+  if (!lastItem && fallbackPos == undefined) {
+    throw new Error(
+      `tried to insert ${toInsert} as first occurence with no fallback position`
+    );
+  }
+  const lastItemPosition: number = lastItem ? lastItem.getEnd() : fallbackPos;
 
-  // Find the decorator declaration.
-  if (!node) {
+  return new InsertChange(file, lastItemPosition, toInsert);
+}
+
+export function findNodes(
+  node: ts.Node,
+  kind: ts.SyntaxKind,
+  max = Infinity
+): ts.Node[] {
+  if (!node || max == 0) {
     return [];
   }
-  // Get all the children property assignment of object literals.
-  const matchingProperties: ts.ObjectLiteralElement[] = (node as ts.ObjectLiteralExpression).properties
-    .filter(prop => prop.kind == ts.SyntaxKind.PropertyAssignment)
-    // Filter out every fields that's not "metadataField". Also handles string literals
-    // (but not expressions).
-    .filter((prop: ts.PropertyAssignment) => {
-      const name = prop.name;
-      switch (name.kind) {
-        case ts.SyntaxKind.Identifier:
-          return (name as ts.Identifier).getText(source) == metadataField;
-        case ts.SyntaxKind.StringLiteral:
-          return (name as ts.StringLiteral).text == metadataField;
-      }
 
-      return false;
+  const arr: ts.Node[] = [];
+  if (node.kind === kind) {
+    arr.push(node);
+    max--;
+  }
+  if (max > 0) {
+    for (const child of node.getChildren()) {
+      findNodes(child, kind, max).forEach(node => {
+        if (max > 0) {
+          arr.push(node);
+        }
+        max--;
+      });
+
+      if (max <= 0) {
+        break;
+      }
+    }
+  }
+
+  return arr;
+}
+
+export function getSourceNodes(sourceFile: ts.SourceFile): ts.Node[] {
+  const nodes: ts.Node[] = [sourceFile];
+  const result = [];
+
+  while (nodes.length > 0) {
+    const node = nodes.shift();
+
+    if (node) {
+      result.push(node);
+      if (node.getChildCount(sourceFile) >= 0) {
+        nodes.unshift(...node.getChildren());
+      }
+    }
+  }
+
+  return result;
+}
+
+export interface Change {
+  apply(host: any): Promise<void>;
+  readonly type: string;
+  readonly path: string | null;
+  readonly order: number;
+  readonly description: string;
+}
+
+export class NoopChange implements Change {
+  type = 'noop';
+  description = 'No operation.';
+  order = Infinity;
+  path = null;
+  apply() {
+    return Promise.resolve();
+  }
+}
+
+export class InsertChange implements Change {
+  type = 'insert';
+  order: number;
+  description: string;
+
+  constructor(public path: string, public pos: number, public toAdd: string) {
+    if (pos < 0) {
+      throw new Error('Negative positions are invalid');
+    }
+    this.description = `Inserted ${toAdd} into position ${pos} of ${path}`;
+    this.order = pos;
+  }
+
+  apply(host: any) {
+    return host.read(this.path).then(content => {
+      const prefix = content.substring(0, this.pos);
+      const suffix = content.substring(this.pos);
+
+      return host.write(this.path, `${prefix}${this.toAdd}${suffix}`);
     });
-
-  // Get the last node of the array literal.
-  if (!matchingProperties) {
-    return [];
   }
-  if (matchingProperties.length == 0) {
-    // We haven't found the field in the metadata declaration. Insert a new field.
-    const expr = node as ts.ObjectLiteralExpression;
-    let position: number;
-    let toInsert: string;
-    if (expr.properties.length == 0) {
-      position = expr.getEnd() - 1;
-      toInsert = `  ${metadataField}: [${expression}]\n`;
-    } else {
-      node = expr.properties[expr.properties.length - 1];
-      position = node.getEnd();
-      // Get the indentation of the last element, if any.
-      const text = node.getFullText(source);
-      if (text.match('^\r?\r?\n')) {
-        toInsert = `,${
-          text.match(/^\r?\n\s+/)[0]
-        }${metadataField}: [${expression}]`;
-      } else {
-        toInsert = `, ${metadataField}: [${expression}]`;
+}
+
+export class RemoveChange implements Change {
+  type = 'remove';
+  order: number;
+  description: string;
+
+  constructor(
+    public path: string,
+    private pos: number,
+    private toRemove: string
+  ) {
+    if (pos < 0) {
+      throw new Error('Negative positions are invalid');
+    }
+    this.description = `Removed ${toRemove} into position ${pos} of ${path}`;
+    this.order = pos;
+  }
+
+  apply(host: any): Promise<void> {
+    return host.read(this.path).then(content => {
+      const prefix = content.substring(0, this.pos);
+      const suffix = content.substring(this.pos + this.toRemove.length);
+      return host.write(this.path, `${prefix}${suffix}`);
+    });
+  }
+}
+
+export class ReplaceChange implements Change {
+  type = 'replace';
+  order: number;
+  description: string;
+
+  constructor(
+    public path: string,
+    private pos: number,
+    private oldText: string,
+    private newText: string
+  ) {
+    if (pos < 0) {
+      throw new Error('Negative positions are invalid');
+    }
+    this.description = `Replaced ${oldText} into position ${pos} of ${path} with ${newText}`;
+    this.order = pos;
+  }
+
+  apply(host: any): Promise<void> {
+    return host.read(this.path).then(content => {
+      const prefix = content.substring(0, this.pos);
+      const suffix = content.substring(this.pos + this.oldText.length);
+      const text = content.substring(this.pos, this.pos + this.oldText.length);
+      if (text !== this.oldText) {
+        return Promise.reject(
+          new Error(`Invalid replace: "${text}" != "${this.oldText}".`)
+        );
       }
-    }
-    const newMetadataProperty = new InsertChange(
-      ngModulePath,
-      position,
-      toInsert
-    );
-    return [newMetadataProperty];
+      return host.write(this.path, `${prefix}${this.newText}${suffix}`);
+    });
   }
-
-  const assignment = matchingProperties[0] as ts.PropertyAssignment;
-
-  // If it's not an array, nothing we can do really.
-  if (assignment.initializer.kind !== ts.SyntaxKind.ArrayLiteralExpression) {
-    return [];
-  }
-
-  const arrLiteral = assignment.initializer as ts.ArrayLiteralExpression;
-  if (arrLiteral.elements.length == 0) {
-    // Forward the property.
-    node = arrLiteral;
-  } else {
-    node = arrLiteral.elements;
-  }
-
-  if (!node) {
-    console.log(
-      'No app module found. Please add your new class to your component.'
-    );
-
-    return [];
-  }
-
-  if (Array.isArray(node)) {
-    const nodeArray = (node as {}) as Array<ts.Node>;
-    const symbolsArray = nodeArray.map(node => node.getText());
-    if (symbolsArray.includes(expression)) {
-      return [];
-    }
-
-    node = node[node.length - 1];
-  }
-
-  let toInsert: string;
-  let position = node.getEnd();
-  if (node.kind == ts.SyntaxKind.ObjectLiteralExpression) {
-    // We haven't found the field in the metadata declaration. Insert a new
-    // field.
-    const expr = node as ts.ObjectLiteralExpression;
-    if (expr.properties.length == 0) {
-      position = expr.getEnd() - 1;
-      toInsert = `  ${metadataField}: [${expression}]\n`;
-    } else {
-      node = expr.properties[expr.properties.length - 1];
-      position = node.getEnd();
-      // Get the indentation of the last element, if any.
-      const text = node.getFullText(source);
-      if (text.match('^\r?\r?\n')) {
-        toInsert = `,${
-          text.match(/^\r?\n\s+/)[0]
-        }${metadataField}: [${expression}]`;
-      } else {
-        toInsert = `, ${metadataField}: [${expression}]`;
-      }
-    }
-  } else if (node.kind == ts.SyntaxKind.ArrayLiteralExpression) {
-    // We found the field but it's empty. Insert it just before the `]`.
-    position--;
-    toInsert = `${expression}`;
-  } else {
-    // Get the indentation of the last element, if any.
-    const text = node.getFullText(source);
-    if (text.match(/^\r?\n/)) {
-      toInsert = `,${text.match(/^\r?\n(\r?)\s+/)[0]}${expression}`;
-    } else {
-      toInsert = `, ${expression}`;
-    }
-  }
-  const insert = new InsertChange(ngModulePath, position, toInsert);
-  return [insert];
 }
 
 export function addParameterToConstructor(
@@ -223,39 +229,6 @@ ${opts.methodHeader} {}
 `;
 
   return [new InsertChange(modulePath, clazz.end - 1, offset(body, 1, true))];
-}
-
-export function removeFromNgModule(
-  source: ts.SourceFile,
-  modulePath: string,
-  property: string
-): Change[] {
-  const nodes = getDecoratorMetadata(source, 'NgModule', '@angular/core');
-  let node: any = nodes[0]; // tslint:disable-line:no-any
-
-  // Find the decorator declaration.
-  if (!node) {
-    return [];
-  }
-
-  // Get all the children property assignment of object literals.
-  const matchingProperty = getMatchingProperty(
-    source,
-    property,
-    'NgModule',
-    '@angular/core'
-  );
-  if (matchingProperty) {
-    return [
-      new RemoveChange(
-        modulePath,
-        matchingProperty.getStart(source),
-        matchingProperty.getFullText(source)
-      )
-    ];
-  } else {
-    return [];
-  }
 }
 
 export function findClass(
@@ -300,128 +273,6 @@ export function offset(
   return wrap ? `\n${lines}\n` : lines;
 }
 
-export function addImportToModule(
-  source: ts.SourceFile,
-  modulePath: string,
-  symbolName: string
-): Change[] {
-  return _addSymbolToNgModuleMetadata(
-    source,
-    modulePath,
-    'imports',
-    symbolName
-  );
-}
-
-export function addImportToTestBed(
-  source: ts.SourceFile,
-  specPath: string,
-  symbolName: string
-): Change[] {
-  const allCalls: ts.CallExpression[] = <any>(
-    findNodes(source, ts.SyntaxKind.CallExpression)
-  );
-
-  const configureTestingModuleObjectLiterals = allCalls
-    .filter(c => c.expression.kind === ts.SyntaxKind.PropertyAccessExpression)
-    .filter(
-      (c: any) => c.expression.name.getText(source) === 'configureTestingModule'
-    )
-    .map(c =>
-      c.arguments[0].kind === ts.SyntaxKind.ObjectLiteralExpression
-        ? c.arguments[0]
-        : null
-    );
-
-  if (configureTestingModuleObjectLiterals.length > 0) {
-    const startPosition = configureTestingModuleObjectLiterals[0]
-      .getFirstToken(source)
-      .getEnd();
-    return [
-      new InsertChange(specPath, startPosition, `imports: [${symbolName}], `)
-    ];
-  } else {
-    return [];
-  }
-}
-
-export function getBootstrapComponent(
-  source: ts.SourceFile,
-  moduleClassName: string
-): string {
-  const bootstrap = getMatchingProperty(
-    source,
-    'bootstrap',
-    'NgModule',
-    '@angular/core'
-  );
-  if (!bootstrap) {
-    throw new Error(`Cannot find bootstrap components in '${moduleClassName}'`);
-  }
-  const c = bootstrap.getChildren();
-  const nodes = c[c.length - 1].getChildren();
-
-  const bootstrapComponent = nodes.slice(1, nodes.length - 1)[0];
-  if (!bootstrapComponent) {
-    throw new Error(`Cannot find bootstrap components in '${moduleClassName}'`);
-  }
-
-  return bootstrapComponent.getText();
-}
-
-function getMatchingObjectLiteralElement(
-  node: any,
-  source: ts.SourceFile,
-  property: string
-) {
-  return (
-    (node as ts.ObjectLiteralExpression).properties
-      .filter(prop => prop.kind == ts.SyntaxKind.PropertyAssignment)
-      // Filter out every fields that's not "metadataField". Also handles string literals
-      // (but not expressions).
-      .filter((prop: ts.PropertyAssignment) => {
-        const name = prop.name;
-        switch (name.kind) {
-          case ts.SyntaxKind.Identifier:
-            return (name as ts.Identifier).getText(source) === property;
-          case ts.SyntaxKind.StringLiteral:
-            return (name as ts.StringLiteral).text === property;
-        }
-        return false;
-      })[0]
-  );
-}
-
-function getMatchingProperty(
-  source: ts.SourceFile,
-  property: string,
-  identifier: string,
-  module: string
-): ts.ObjectLiteralElement {
-  const nodes = getDecoratorMetadata(source, identifier, module);
-  let node: any = nodes[0]; // tslint:disable-line:no-any
-
-  if (!node) return null;
-
-  // Get all the children property assignment of object literals.
-  return getMatchingObjectLiteralElement(node, source, property);
-}
-
-export function addRoute(
-  ngModulePath: string,
-  source: ts.SourceFile,
-  route: string
-): Change[] {
-  const routes = getListOfRoutes(source);
-  if (!routes) return [];
-
-  if (routes.hasTrailingComma || routes.length === 0) {
-    return [new InsertChange(ngModulePath, routes.end, route)];
-  } else {
-    return [new InsertChange(ngModulePath, routes.end, `, ${route}`)];
-  }
-}
-
 export function addIncludeToTsConfig(
   tsConfigPath: string,
   source: ts.SourceFile,
@@ -434,37 +285,6 @@ export function addIncludeToTsConfig(
   } else {
     return [];
   }
-}
-
-function getListOfRoutes(source: ts.SourceFile): ts.NodeArray<ts.Expression> {
-  const imports: any = getMatchingProperty(
-    source,
-    'imports',
-    'NgModule',
-    '@angular/core'
-  );
-
-  if (imports.initializer.kind === ts.SyntaxKind.ArrayLiteralExpression) {
-    const a = imports.initializer as ts.ArrayLiteralExpression;
-
-    for (let e of a.elements) {
-      if (e.kind === ts.SyntaxKind.CallExpression) {
-        const ee = e as ts.CallExpression;
-        const text = ee.expression.getText(source);
-        if (
-          (text === 'RouterModule.forRoot' ||
-            text === 'RouterModule.forChild') &&
-          ee.arguments.length > 0
-        ) {
-          const routes = ee.arguments[0];
-          if (routes.kind === ts.SyntaxKind.ArrayLiteralExpression) {
-            return (routes as ts.ArrayLiteralExpression).elements;
-          }
-        }
-      }
-    }
-  }
-  return null;
 }
 
 export function getImport(
@@ -490,45 +310,6 @@ export function getImport(
   });
 }
 
-export function addProviderToModule(
-  source: ts.SourceFile,
-  modulePath: string,
-  symbolName: string
-): Change[] {
-  return _addSymbolToNgModuleMetadata(
-    source,
-    modulePath,
-    'providers',
-    symbolName
-  );
-}
-
-export function addDeclarationToModule(
-  source: ts.SourceFile,
-  modulePath: string,
-  symbolName: string
-): Change[] {
-  return _addSymbolToNgModuleMetadata(
-    source,
-    modulePath,
-    'declarations',
-    symbolName
-  );
-}
-
-export function addEntryComponents(
-  source: ts.SourceFile,
-  modulePath: string,
-  symbolName: string
-): Change[] {
-  return _addSymbolToNgModuleMetadata(
-    source,
-    modulePath,
-    'entryComponents',
-    symbolName
-  );
-}
-
 export function addGlobal(
   source: ts.SourceFile,
   modulePath: string,
@@ -545,24 +326,24 @@ export function addGlobal(
   }
 }
 
-export function insert(host: Tree, modulePath: string, changes: Change[]) {
+export function insert(host: Tree, modulePath: string, changes: any[]) {
   if (changes.length < 1) {
     return;
   }
   const recorder = host.beginUpdate(modulePath);
   for (const change of changes) {
-    if (change instanceof InsertChange) {
+    if (change.type === 'insert') {
       recorder.insertLeft(change.pos, change.toAdd);
-    } else if (change instanceof RemoveChange) {
+    } else if (change.type === 'remove') {
       recorder.remove((<any>change).pos - 1, (<any>change).toRemove.length + 1);
-    } else if (change instanceof NoopChange) {
+    } else if (change.type === 'noop') {
       // do nothing
-    } else if (change instanceof ReplaceChange) {
+    } else if (change.type === 'replace') {
       const action = <any>change;
       recorder.remove(action.pos, action.oldText.length);
       recorder.insertLeft(action.pos, action.newText);
     } else {
-      throw new Error(`Unexpected Change '${change}'`);
+      throw new Error(`Unexpected Change '${change.constructor.name}'`);
     }
   }
   host.commitUpdate(recorder);
@@ -637,210 +418,6 @@ export function getProjectConfig(host: Tree, name: string): any {
   } else {
     return projectConfig;
   }
-}
-
-export function updateProjectConfig(name: string, projectConfig: any): Rule {
-  return updateJsonInTree('/angular.json', angularJson => {
-    angularJson.projects[name] = projectConfig;
-    return angularJson;
-  });
-}
-
-export function readBootstrapInfo(
-  host: Tree,
-  app: string
-): {
-  moduleSpec: string;
-  modulePath: string;
-  mainPath: string;
-  moduleClassName: string;
-  moduleSource: ts.SourceFile;
-  bootstrapComponentClassName: string;
-  bootstrapComponentFileName: string;
-} {
-  const config = getProjectConfig(host, app);
-
-  let mainPath;
-  try {
-    mainPath = config.architect.build.options.main;
-  } catch (e) {
-    throw new Error('Main file cannot be located');
-  }
-
-  if (!host.exists(mainPath)) {
-    throw new Error('Main file cannot be located');
-  }
-
-  const mainSource = host.read(mainPath)!.toString('utf-8');
-  const main = ts.createSourceFile(
-    mainPath,
-    mainSource,
-    ts.ScriptTarget.Latest,
-    true
-  );
-  const moduleImports = getImport(
-    main,
-    (s: string) => s.indexOf('.module') > -1
-  );
-  if (moduleImports.length !== 1) {
-    throw new Error(`main.ts can only import a single module`);
-  }
-  const moduleImport = moduleImports[0];
-  const moduleClassName = moduleImport.bindings.filter(b =>
-    b.endsWith('Module')
-  )[0];
-
-  const modulePath = `${path.join(
-    path.dirname(mainPath),
-    moduleImport.moduleSpec
-  )}.ts`;
-  if (!host.exists(modulePath)) {
-    throw new Error(`Cannot find '${modulePath}'`);
-  }
-
-  const moduleSourceText = host.read(modulePath)!.toString('utf-8');
-  const moduleSource = ts.createSourceFile(
-    modulePath,
-    moduleSourceText,
-    ts.ScriptTarget.Latest,
-    true
-  );
-
-  const bootstrapComponentClassName = getBootstrapComponent(
-    moduleSource,
-    moduleClassName
-  );
-  const bootstrapComponentFileName = `./${path.join(
-    path.dirname(moduleImport.moduleSpec),
-    `${toFileName(
-      bootstrapComponentClassName.substring(
-        0,
-        bootstrapComponentClassName.length - 9
-      )
-    )}.component`
-  )}`;
-
-  return {
-    moduleSpec: moduleImport.moduleSpec,
-    mainPath,
-    modulePath,
-    moduleSource,
-    moduleClassName,
-    bootstrapComponentClassName,
-    bootstrapComponentFileName
-  };
-}
-
-export function addClass(
-  source: ts.SourceFile,
-  modulePath: string,
-  clazzName: string,
-  clazzSrc: string
-): Change {
-  if (!findClass(source, clazzName, true)) {
-    const nodes = findNodes(source, ts.SyntaxKind.ClassDeclaration);
-    return insertAfterLastOccurrence(
-      nodes,
-      offset(clazzSrc, 0, true),
-      modulePath,
-      0,
-      ts.SyntaxKind.ClassDeclaration
-    );
-  }
-  return new NoopChange();
-}
-
-/**
- * e.g
- * ```ts
- *   export type <Feature>Actions = <Feature> | Load<Feature>s | <Feature>sLoaded | <Feature>sLoadError;
- * ```
- */
-export function addUnionTypes(
-  source: ts.SourceFile,
-  modulePath: string,
-  typeName: string,
-  typeValues: string[]
-) {
-  const target: ts.TypeAliasDeclaration = findNodesOfType(
-    source,
-    ts.SyntaxKind.TypeAliasDeclaration,
-    it => it.name.getText() === typeName
-  );
-  if (!target) {
-    throw new Error(`Cannot find union type '${typeName}'`);
-  }
-
-  const node = target.type as ts.TypeReferenceNode;
-
-  // Append new types to create a union type...
-  return new InsertChange(
-    modulePath,
-    node.end,
-    ['', ...typeValues].join(' | ')
-  );
-}
-
-/**
- * Add 1..n enumerators using name + (optional) value pairs
- */
-export function addEnumeratorValues(
-  source: ts.SourceFile,
-  modulePath: string,
-  enumName: string,
-  pairs: NameValue[] = []
-): Change[] {
-  const target = findNodesOfType(
-    source,
-    ts.SyntaxKind.EnumDeclaration,
-    it => it.name.getText() === enumName
-  );
-  const list = target ? target.members : undefined;
-  if (!target) {
-    throw new Error(`Cannot find enum '${enumName}'`);
-  }
-  const addComma = !(list.hasTrailingComma || list.length === 0);
-
-  return pairs.reduce((buffer, it) => {
-    const member = it.value ? `${it.name} = '${it.value}'` : it.name;
-    const memberExists = () => {
-      return list.filter(m => m.name.getText() === it.name).length;
-    };
-
-    if (memberExists()) {
-      throw new Error(`Enum '${enumName}.${it.name}' already exists`);
-    }
-
-    return [
-      ...buffer,
-      new InsertChange(modulePath, list.end, (addComma ? ', ' : '') + member)
-    ];
-  }, []);
-}
-
-/**
- * Find Enum declaration in source based on name
- * e.g.
- *    export enum ProductsActionTypes {
- *       ProductsAction = '[Products] Action'
- *    }
- */
-const IDENTITY = a => a;
-export function findNodesOfType(
-  source: ts.Node,
-  kind: ts.SyntaxKind,
-  predicate: (a: any) => boolean,
-  extract: (a: any) => any = IDENTITY,
-  firstOnly: boolean = true
-): any {
-  const nodes = findNodes(source, kind);
-  const matching = nodes.filter((i: any) => predicate(i)).map(extract);
-  return matching.length ? (firstOnly ? matching[0] : matching) : undefined;
-}
-
-export interface NameValue {
-  name: string;
-  value?: string;
 }
 
 export function createOrUpdate(host: Tree, path: string, content: string) {
@@ -939,30 +516,6 @@ export function insertImport(
     fallbackPos,
     ts.SyntaxKind.StringLiteral
   );
-}
-
-export function getDecoratorPropertyValueNode(
-  host: Tree,
-  modulePath: string,
-  identifier: string,
-  property: string,
-  module: string
-) {
-  const moduleSourceText = host.read(modulePath)!.toString('utf-8');
-  const moduleSource = ts.createSourceFile(
-    modulePath,
-    moduleSourceText,
-    ts.ScriptTarget.Latest,
-    true
-  );
-  const templateNode = getMatchingProperty(
-    moduleSource,
-    property,
-    identifier,
-    module
-  );
-
-  return templateNode.getChildAt(templateNode.getChildCount() - 1);
 }
 
 export function replaceNodeValue(
